requirements.txt
# Core
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
rich>=13.7.0
click>=8.1.7
pathlib2>=2.3.7

# Prediction Market APIs
py-clob-client>=0.17.0
kalshi-python-sync>=2.0.0

# AI / Anthropic
anthropic>=0.40.0

# ML / RL
xgboost>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
joblib>=1.3.0

# MCP integration
mcp>=1.0.0

# Database
# sqlite3 is built into Python stdlib

# Alerts
# telegram via requests (no heavy SDK needed)

# Dev / testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
